requirements.txt
streamlit
flask
requests
python-dotenv
pandas
numpy
pymongo
pdfplumber
python-docx